Move graphql_controller out of api v1
